refactored the login form code
