+++ .env.development
@@ -1,2 +1,3 @@
 REACT_APP_BACKEND_URI='http://localhost:4001'
 REACT_APP_SENTRY_DSN=https://814d455c01b34f7885485a6f86bf745d@sentry.io/1378772
+SKIP_PREFLIGHT_CHECK=true
